restore avatar colors move files
